more security measures
Essentially if /proc/cmdline is missing "boot" in as it's argument it's
obvious somebody is trying to boot on the wrong device as you'd have
GRUB or alike have that argument.

Furthermore the fscrypt will only fail if either of the UUID's have
changed, aka somebody cloned the drive or somebody is trying to boot on
a different system, really cool stuff.

And as a haha funny, I made it just sit there forever never to do
anything, with the while true loop. I would've wanted it to echo
something on screen but I don't know if that's possible.