turns out i can show stuff on screen
Apparently the reason they weren't working before was because I used the
argument "console=tty2" to try and make the screen less verbose. And all
output is displayed on tty1, hence why changing the tty decreasing the
amount of things on screen.

I don't know any way I can switch the tty while it's booting in the
early stages, as chvt didn't seem to work for whatever reason. And I'm
not assed to discover it any further. But if the "boot" argument is
missing and that's why it can't boot further then you probably also
wouldn't have the "console=tty2" argument, therefore you would see the
message I wrote.